fix: 초기 sort 제대로하게 바꿈
