+++ src/rank/components/chatSummary.jsx
@@ -6,29 +6,7 @@ import { WeightContext } from "../../contexts/weightProvider";
 import { getIndicatorsResult } from "../services/indicators";
 
 export default function ChatSummary({ userId, name }) {
-  const { sliderValues, setSliderValues } = useContext(WeightContext);
-  const rates = ["안정성", "수익성", "활동성", "성장성", "언급량", "감정지수"];
-  const [maxPreference, setMaxPreference] = useState("");
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const values = await getIndicatorsResult();
-      const valuesToRates = [];
-      valuesToRates.push(values.수익성);
-      valuesToRates.push(values.안전성);
-      valuesToRates.push(values.성장성);
-      valuesToRates.push(values.활동성);
-      valuesToRates.push(values.언급량);
-      valuesToRates.push(values.감성지수);
-      setSliderValues(valuesToRates);
-      console.log(sliderValues);
-      const maxIdx = valuesToRates.reduce((maxIndex, currentValue, currentIndex, array) => {
-        return currentValue > array[maxIndex] ? currentIndex : maxIndex;
-      }, 0);
-      setMaxPreference(rates[maxIdx]);
-    };
-    fetchData();
-  }, []);
+  const { sliderValues, setSliderValues, highest } = useContext(WeightContext);
 
   return (
     <div className="chatSummaryBallon">
@@ -38,7 +16,7 @@ export default function ChatSummary({ userId, name }) {
           {name}님에게 적합한 주식 종목 순위를 보여드릴게요.{" "}
         </p>
         <p style={{ color: CHAT_SAY }}>
-          대화 결과, {name}님은 {maxPreference}이 높은 주식을 선호하는 것 같아요. 만약 다른 결과를
+          대화 결과, {name}님은 {highest}이 높은 주식을 선호하는 것 같아요. 만약 다른 결과를
           보고싶다면 아래의 슬라이더를 조절하거나 다시 대화를 시작해보세요. 종목을 클릭하면 자세한
           정보를 볼 수 있고, 다이렉트 인덱싱을 위해 장바구니에 담을 수 있어요.
         </p>
